+++ src/main/java/cart/ui/common/ControllerExceptionHandler.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.ui.common;
 
 import static cart.exception.ErrorMessage.INTERNAL_SERVER_ERROR;
